+++ test/api/test_sonarr.py
@@ -352,6 +352,7 @@ def test_get_queue(mock_get, cli):
     mock_get.assert_called_with(cli.api_url_queue, url_params=data)
     assert res == TEST_JSON
 
+
 @patch("pycliarr.api.sonarr.BaseCliMediaApi.request_get", return_value=TEST_JSON)
 def test_get_queue_with_args(mock_get, cli):
     res = cli.get_queue(page = 2, sort_key = "sort", page_size = 3, sort_dir = "asc", include_unknown = False)
@@ -366,8 +367,15 @@ def test_get_queue_with_args(mock_get, cli):
     mock_get.assert_called_with(cli.api_url_queue, url_params=data)
     assert res == TEST_JSON
 
+
 @patch("pycliarr.api.base_media.BaseCliApi.request_post", return_value=TEST_JSON)
 def test_renamefiles(mock_base, cli):
     res = cli.rename_files([1, 2, 3], 1)
     mock_base.assert_called_with(cli.api_url_command, json_data={"name": "RenameFiles", "files": [1, 2, 3], "seriesId": 1})
     assert res == TEST_JSON
+
+
+@patch("pycliarr.api.sonarr.BaseCliMediaApi.request_get")
+def test_get_rename_serie(mock_base, cli):
+    res = cli.get_rename(1234)
+    mock_base.assert_called_with(cli.api_url_rename, url_params={"seriesId": 1234})